:building_construction: moving implementation to .cpp file
