+++ doc/Doxyfile.in
@@ -113,8 +113,8 @@ WARN_LOGFILE           =
 #---------------------------------------------------------------------------
 # Configuration options related to the input files
 #---------------------------------------------------------------------------
-INPUT                  = @CMAKE_CURRENT_SOURCE_DIR@/include/crow \
-                         @CMAKE_CURRENT_SOURCE_DIR@/README.md
+INPUT                  = @CMAKE_CURRENT_SOURCE_DIR@/include/crow @CMAKE_CURRENT_SOURCE_DIR@/src \
+                         @CMAKE_CURRENT_BINARY_DIR@/include/crow @CMAKE_CURRENT_SOURCE_DIR@/README.md
 INPUT_ENCODING         = UTF-8
 FILE_PATTERNS          = *.c \
                          *.cc \
