+++ include/crow/config.h.in
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef NLOHMANN_CROW_CONFIG
+#define NLOHMANN_CROW_CONFIG
 
 #cmakedefine NLOHMANN_CROW_HAVE_CXXABI_H
 #cmakedefine NLOHMANN_CROW_HAVE_EXECINFO_H
@@ -24,3 +25,5 @@
 #define NLOHMANN_CROW_VERSION_MINOR ${PROJECT_VERSION_MINOR}
 #define NLOHMANN_CROW_VERSION_PATCH ${PROJECT_VERSION_PATCH}
 #define NLOHMANN_CROW_VERSION "${PROJECT_VERSION}"
+
+#endif
