+++ maintainer/Makefile
@@ -9,7 +9,7 @@ pretty:
 	   --indent-col1-comments --pad-oper --pad-header --align-pointer=type \
 	   --align-reference=type --add-brackets --convert-tabs --close-templates \
 	   --lineend=linux --preserve-date --suffix=none --formatted --recursive \
-	   "include/crow/*.hpp,cpp" "tests/*.cpp,hpp"
+	   "include/crow/*.hpp,cpp" "tests/*.cpp,hpp" "src/*.cpp,*.hpp"
 
 upload_docs:
 	mkdir build ; cd build ; cmake .. ; cmake --build . --target crow_documentation ; cd ..
